apply suggestions of pr
