+++ frontend/src/services/errorConstants.ts
@@ -1,12 +1,13 @@
 export enum ErrorConstants {
   VALIDATION_ERROR = "VALIDATION_ERROR",
   UNIQUE_ERROR = "UNIQUE_ERROR",
-  END_OF_REGISTRATION_ERROR = "END_OF_REGISTRATION_ERROR",
   BAD_REQUEST_ERROR = "BAD_REQUEST_ERROR",
   ACCESS_DENIED_ERROR = "ACCESS_DENIED_ERROR",
   RESOURCE_ALREADY_EXISTS_ERROR = "RESOURCE_ALREADY_EXISTS_ERROR",
   EXISTING_DEPENDENCY_ERROR = "EXISTING_DEPENDENCY_ERROR",
   NOT_FOUND_ERROR = "NOT_FOUND_ERROR",
+  WRONG_TIME_EXCEPTION = "WRONG_TIME_EXCEPTION",
+  NOT_DELETABLE_ERROR = "NOT_DELETABLE_ERROR",
 }
 
 export interface ErrorResponse {
