+++ frontend/src/views/PcrTestView.vue
@@ -26,7 +26,7 @@
       />
 
       <!-- PCR TEST ID DOES EXIST  -->
-      <v-row v-if="!loading && !error && isInDateRange">
+      <v-row v-if="!error && isInDateRange">
         <h1 class="header mb-8">Teilnehmer zum PCR Pool hinzuzufügen.</h1>
         <v-col justify="center" align="center">
           <div class="mb-8">
@@ -47,57 +47,11 @@
             @submitCode="addAttendeeToPcrPool"
           />
 
-          <v-row ref="attendeeListRef">
-            <v-row justify="center" v-if="attendees.length > 0">
-              <v-list subheader two-line class="attendee-list">
-                <v-subheader inset>Teilnehmer</v-subheader>
-
-                <v-list-item
-                  v-for="attendee in attendees"
-                  :key="attendee.attendeeCode"
-                >
-                  <v-list-item-avatar>
-                    <v-icon class="grey lighten-1" dark> mdi-account </v-icon>
-                  </v-list-item-avatar>
-
-                  <v-list-item-content>
-                    <v-list-item-title
-                      v-text="
-                        `${attendee.attendeeFirstName} ${attendee.attendeeLastName}`
-                      "
-                    ></v-list-item-title>
-
-                    <v-list-item-subtitle
-                      v-text="attendee.departmentName"
-                    ></v-list-item-subtitle>
-                  </v-list-item-content>
-
-                  <v-list-item-action>
-                    <v-btn
-                      icon
-                      v-if="isInDateRange"
-                      @click="removeAttendeeFromPcrPool(attendee.attendeeCode)"
-                    >
-                      <v-icon
-                        color="grey lighten-1"
-                        @mouseover="trashIndex = attendee.attendeeCode"
-                        v-show="trashIndex !== attendee.attendeeCode"
-                      >
-                        mdi-delete
-                      </v-icon>
-                      <v-icon
-                        color="grey lighten-1"
-                        @mouseleave="trashIndex = ''"
-                        v-show="trashIndex === attendee.attendeeCode"
-                      >
-                        mdi-delete-empty
-                      </v-icon>
-                    </v-btn>
-                  </v-list-item-action>
-                </v-list-item>
-              </v-list>
-            </v-row>
-          </v-row>
+          <AttendeeList
+            ref="attendeeListRef"
+            :attendees="attendees"
+            @refetch="refetchData"
+          />
         </v-col>
       </v-row>
     </v-container>
@@ -107,25 +61,27 @@
 <script lang="ts">
 import { Vue, Component } from "vue-property-decorator";
 import Scanner from "../components/Scanner.vue";
+import Sorry from "../components/Sorry.vue";
+import AttendeeList from "../components/AttendeeList.vue";
 import {
   getPcrPool,
   addAttendeeToPcrPool,
-  removeAttendeeFromPool,
   PcrAttendee,
   PcrTest,
 } from "../services/pcrTest";
-import Sorry from "../components/Sorry.vue";
 import {
   ErrorConstants,
   ErrorResponse,
   isErrorOfType,
 } from "../services/errorConstants";
 
-@Component({ name: "PcrTestView", components: { Scanner, Sorry } })
+@Component({
+  name: "PcrTestView",
+  components: { Scanner, Sorry, AttendeeList },
+})
 export default class PcrTestView extends Vue {
   private pcrPoolId: string = "";
   private pcrTest: PcrTest | null = null;
-  private trashIndex = "";
   private error: ErrorResponse | null = null;
   private loading: boolean = false;
   $refs!: {
@@ -171,24 +127,6 @@ export default class PcrTestView extends Vue {
       this.$vuetify.goTo(this.$refs.attendeeListRef, {
         easing: "easeInOutCubic",
       });
-      await this.refetchData();
-    }
-  }
-
-  protected async removeAttendeeFromPcrPool(
-    attendeeCode: string
-  ): Promise<void> {
-    const res = await removeAttendeeFromPool(this.pcrPoolId, attendeeCode);
-
-    if (res?.ok) {
-      const deletedAttendees = this.attendees.splice(
-        this.attendees.findIndex((v) => v.attendeeCode === attendeeCode),
-        1
-      );
-      this.$toast.info(
-        `${deletedAttendees[0].attendeeFirstName} ${deletedAttendees[0].attendeeLastName} wurde entfernt.`
-      );
-
       await this.refetchData();
     }
   }
